Refactoring app structure to account for Crew and Resident views
