--- components/Navbar.js
@@ -1,1 +1,0 @@
-
