+++ components/resident/AddPotholeLocation.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Platform, StyleSheet, Dimensions } from 'react-native';
 import { MapView, Constants, Location, Permissions } from 'expo';
-import { getGeocodedAddress, fetchPotholes } from '../store/potholes';
-import { createUpdateLocationAction } from '../store/report';
+import { getGeocodedAddress, fetchPotholes } from '../../store/potholes';
+import { createUpdateLocationAction } from '../../store/report';
 import { Container, Content, Text, Card, Form, Item,
   Input,
   Button,
